requirements.txt
Django==1.6.6
Pillow==2.5.3
argparse==1.2.1
djangorestframework==2.4.1
wsgiref==0.1.2
